Alteração no frontend do programa + criação do arquivo views/layout.php + alteração dos demais arquivos desta  pasta + criação do arquivo assets/style.css + alteração no arquivo Controller.php na pasta Core

+++ api/views/product/index.php
@@ -1,38 +1,36 @@
-<main>
-  <div class="container">
-    <div class="row">
-      <div class="col-8 offset-2" style="margin-top:30px">      
-          <div class="d-flex justify-content-between align-items-center mb-3">
-              <h2>Produtos</h2> 
-              <a href="<?= BASE_URL ?>/product/create" class="btn btn-success"> Incluir Produto </a>
-              <a href="<?= BASE_URL ?>" class="btn btn-primary"> Página Principal </a>
-          </div>
+<div class="dashboard-box">
+  <div class="d-flex justify-content-between align-items-center mb-4">
 
-        <table class="table">
-          <thead>
-            <tr>
-              <th scope="col">ID</th>
-              <th scope="col">Nome</th>
-              <th colspan="3">Ações</th>
-            </tr>
-          </thead>
-          <tbody>
-            <?php foreach ($data['products'] as $product) { ?>
-            <tr>
-              <td><?= $product['id'] ?></td>
-              <td><?= $product['nome'] ?></td>
-
-              <td> <a href="<?= BASE_URL ?>/product/show/<?= $product['id'] ?>"> Visualizar </a></td>
-              <td> <a href="<?= BASE_URL ?>/product/edit/<?= $product['id'] ?>">Editar</a></td>
-              <td> <a href="<?= BASE_URL ?>/product/delete/<?= $product['id'] ?>"
-                onclick="return confirm('Deseja excluir este produto?')">Excluir
-                </a>
-              </td>
-            </tr>
-            <?php }?>
-          </tbody>
-        </table>
-      </div>
+    <h2><i class="bi bi-box-seam"> </i> - Produtos</h2>
+    <div>
+      <a href="<?= BASE_URL ?>/product/create" class="btn btn-success"> Incluir Produto</a>
     </div>
+
   </div>
-</main>
+
+  <div class="table-scroll">
+  <table class="table">
+    <thead>
+      <tr>
+        <th>ID</th>
+        <th>Nome</th>
+        <th colspan="3">Ações</th>
+      </tr>
+
+    </thead>
+    <tbody>
+      <?php foreach ($data['products'] as $product) { ?>
+
+      <tr>
+        <td><?= $product['id'] ?></td>
+        <td><?= $product['nome'] ?></td>
+        <td><a href="<?= BASE_URL ?>/product/show/<?= $product['id'] ?>">Visualizar</a></td>
+        <td><a href="<?= BASE_URL ?>/product/edit/<?= $product['id'] ?>">Editar</a></td>
+        <td><a href="<?= BASE_URL ?>/product/delete/<?= $product['id'] ?>"onclick="return confirm('Deseja excluir este produto?')">Excluir</a></td>
+      </tr>
+
+      <?php } ?>
+    </tbody>
+  </table>
+</div>
+</div>
